# lib/api-zod/src/generated/types/postInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for LNCT GenZ Connect
 * OpenAPI spec version: 0.1.0
 */
import type { CategorySlug } from './categorySlug';
import type { PostInputExpiryDuration } from './postInputExpiryDuration';
import type { PostInputUrgency } from './postInputUrgency';

export interface PostInput {
  category: CategorySlug;
  /**
     * @minLength 3
     * @maxLength 120
     */
  title: string;
  /**
     * @minLength 10
     * @maxLength 2000
     */
  description: string;
  skills?: string[];
  imageUrl?: string;
  imageUrls?: string[];
  urgency?: PostInputUrgency;
  expiryDuration?: PostInputExpiryDuration;
  /** @maxLength 300 */
  contactNote?: string;
}